Split base design out of `BeatmapAttributesDisplay`
